# [8_12] improve strategy of expanding string for performance 